Refactored progresswindow to ProgressWindow to follow our CamelCase convention #195
